pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.42.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.5",
    "matplotlib>=3.10.5",
    "networkx>=3.5",
    "pydot>=4.0.1",
    "python-dotenv>=1.1.1",
]
